Вынес функции в отдельный файл
